+++ SOS/MIDIEvent.cs
@@ -4,7 +4,7 @@
     {
         public readonly byte velocity;
         public readonly byte note;
-        public MIDIEvent(byte vel, byte no, byte deltatime)
+        public MIDIEvent(byte vel, byte no, int deltatime)
         {
             deltaTime = deltatime;
             eventType = 0;
